Add edge level configuration
It's nice to be able to reduce the number of edges in a graph in a
controlled way. By introcuding `edge_level` which controls the
label of edges, it becomes possible to simplify labels up to the level
when they are empty.

Empty labels will cause duplication in paths and the code already
removes duplicates.

In effect this causes the number of connections to be reduced to a
single connection between nodes. :v: