Include routing keys which match queue names
I was filtering out bindings which had their destinations (queue names)
equal to the routing keys.

This turns out to be a bad idea as out in the wild, nothing prevents
anyone from naming their routing keys the same as their queues.

After looking at the payloads again it turns out I can use the `source`
field to remove anything that is not coming from an exchange, which fits
the original intended purpose.