Assert that a live consumer is mapped
This integration test verifies that

- routing key decomposition works as expected
- consumer tags are mapped as expected
- the underlying code can connect to the API

I needed a way to detect whether the consumer was up and running and I
did not want to do it via the `topology` method as that was the code
under test.

As a result, "consumer readiness" is tested via HTTP call to the
RabbitMQ API rather than any other gem, which means I pulled in the
`rest-client` library.